Better navigation, better file remove, few typo fix

+++ src/main/java/com/spectral369/ARD/AdeverintaRadiereAutoPDF.java
@@ -1,6 +1,5 @@
 package com.spectral369.ARD;
 
-import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -19,6 +18,8 @@ import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.router.AfterNavigationEvent;
 import com.vaadin.flow.router.AfterNavigationObserver;
+import com.vaadin.flow.router.BeforeLeaveEvent;
+import com.vaadin.flow.router.BeforeLeaveObserver;
 import com.vaadin.flow.router.PageTitle;
 import com.vaadin.flow.router.RouteConfiguration;
 import com.vaadin.flow.router.RouterLayout;
@@ -26,7 +27,8 @@ import com.vaadin.flow.router.RouterLink;
 
 @PageTitle("AdeverintaRadiere")
 //@Route("ADVRadiereAutoPDF")//:tm
-public class AdeverintaRadiereAutoPDF extends HorizontalLayout implements RouterLayout, AfterNavigationObserver {
+public class AdeverintaRadiereAutoPDF extends HorizontalLayout
+	implements RouterLayout, AfterNavigationObserver, BeforeLeaveObserver {
     private static final long serialVersionUID = 1L;
     public static final String NAME = "ADVRadiereAutoPDF";
     public static String FNAME;
@@ -39,6 +41,8 @@ public class AdeverintaRadiereAutoPDF extends HorizontalLayout implements Router
     Button backbtn;
     String fileName = null;
     PdfView pdfView = null;
+    String browser = null;
+    private Map<String, List<String>> parameters = null;
 
     static {
 	AdeverintaRadiereAutoPDF.FNAME = "";
@@ -78,24 +82,6 @@ public class AdeverintaRadiereAutoPDF extends HorizontalLayout implements Router
 	RouterLink routerLink = new RouterLink("", MainView.class);
 	routerLink.getElement().appendChild(backbtn.getElement());
 
-	this.backbtn.addClickListener(evt -> {
-
-	    if (fileName != null) {
-
-		File pd = new File(Utils.getFullPath(fileName, false));
-		try {
-		    System.out.println("File Deleted?: " + Files.deleteIfExists(Path.of(pd.getAbsolutePath())));
-		} catch (IOException e) {
-
-		    e.printStackTrace();
-		}
-
-		PdfList.deleteFile(fileName);
-
-	    }
-	    RouteConfiguration.forSessionScope().removeRoute(AdeverintaRadiereAutoPDF.class);
-	    RouteConfiguration.forSessionScope().removeRoute(NAME);
-	});
 	backLayout.add(routerLink);
 	content.add(backLayout);
 	content.setAlignItems(Alignment.CENTER);
@@ -104,9 +90,21 @@ public class AdeverintaRadiereAutoPDF extends HorizontalLayout implements Router
 
 	setSizeFull();
 
-	UI.getCurrent().getPage().executeJs(
-		"window.addEventListener('beforeunload', function (e) {    $0.$server.windowClosed(); var nAgt = navigator.userAgent;if ((verOffset=nAgt.indexOf('Chrome'))!=-1) { (e || window.event).returnValue = null ; } return; });",
-		getElement());
+	/*
+	 * UI.getCurrent().getPage().executeJs(
+	 * "window.addEventListener('beforeunload', function (e) {    $0.$server.windowClosed(); var nAgt = navigator.userAgent;if ((verOffset=nAgt.indexOf('Chrome'))!=-1) { (e || window.event).returnValue = null ; } return; });"
+	 * , getElement()); //triggers on tab close with alert message !!!
+	 */
+	UI.getCurrent().getPage()
+		.executeJs("window.addEventListener('beforeunload', () => $0.$server.windowClosed()); ", getElement()); // does
+															// not
+															// trigger
+															// on
+															// tab
+															// close
+															// !!!!!!!
+	UI.getCurrent().getPage().executeJs("window.addEventListener('unload', () => $0.$server.windowClosed()); ",
+		getElement()); // does trigger on tab close !!!!!!!
 
     }
 
@@ -120,14 +118,15 @@ public class AdeverintaRadiereAutoPDF extends HorizontalLayout implements Router
 
 	    e.printStackTrace();
 	}
-	PdfList.deleteFile(fileName);
+	if (PdfList.isFilePresent(fileName))
+	    PdfList.deleteFile(fileName);
 	RouteConfiguration.forSessionScope().removeRoute(NAME);
 	RouteConfiguration.forSessionScope().removeRoute(AdeverintaRadiereAutoPDF.class);
     }
 
     @Override
     public void afterNavigation(AfterNavigationEvent event) {
-	Map<String, List<String>> parameters = event.getLocation().getQueryParameters().getParameters();
+	parameters = event.getLocation().getQueryParameters().getParameters();
 
 	if (fileName == null) {
 	    fileName = new String(parameters.get("tm").get(0));
@@ -139,6 +138,29 @@ public class AdeverintaRadiereAutoPDF extends HorizontalLayout implements Router
 
 	RouteConfiguration.forSessionScope().removeRoute(AdeverintaRadiereAutoPDF.class);
 	RouteConfiguration.forSessionScope().removeRoute(NAME);
+    }
+
+    @Override
+    public void beforeLeave(BeforeLeaveEvent event) {
+
+	if (fileName == null) {
+	    fileName = new String(parameters.get("tm").get(0));
+	}
+	if (!fileName.isEmpty()) {
+
+	    pdfView.add(Utils.getFullPath(fileName, true));
+	}
+	try {
+	    System.out.println(Files.deleteIfExists(Path.of(Utils.getFullPath(fileName, false))));
+	    if (PdfList.isFilePresent(fileName))
+		PdfList.deleteFile(fileName);
+	} catch (IOException e) {
+
+	    e.printStackTrace();
+	}
+	RouteConfiguration.forSessionScope().removeRoute(AdeverintaRadiereAutoPDF.class);
+	RouteConfiguration.forSessionScope().removeRoute(NAME);
+
     }
 
 }
