+++ src/main/java/com/spectral369/CEAF/CerereEmitereAutorizatieFunctionarePDF.java
@@ -1,13 +1,12 @@
 package com.spectral369.CEAF;
 
-import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.List;
 import java.util.Map;
 
-
+import com.spectral369.ARD.AdeverintaRadiereAutoPDF;
 import com.spectral369.birotica.MainView;
 import com.spectral369.birotica.PdfList;
 import com.spectral369.utils.PdfView;
@@ -20,12 +19,14 @@ import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.router.AfterNavigationEvent;
 import com.vaadin.flow.router.AfterNavigationObserver;
+import com.vaadin.flow.router.BeforeLeaveEvent;
+import com.vaadin.flow.router.BeforeLeaveObserver;
 import com.vaadin.flow.router.RouteConfiguration;
 import com.vaadin.flow.router.RouterLayout;
 import com.vaadin.flow.router.RouterLink;
 
 public class CerereEmitereAutorizatieFunctionarePDF extends HorizontalLayout
-	implements RouterLayout, AfterNavigationObserver {
+	implements RouterLayout, AfterNavigationObserver,BeforeLeaveObserver {
     private static final long serialVersionUID = 1L;
     public static final String NAME = "CerereEmitereAutorizatieFunctionarePDF";
     public static String FNAME;
@@ -38,6 +39,7 @@ public class CerereEmitereAutorizatieFunctionarePDF extends HorizontalLayout
     Button backbtn;
     String fileName = null;
     PdfView pdfView = null;
+    private Map<String, List<String>> parameters = null;
 
     static {
 	CerereEmitereAutorizatieFunctionarePDF.FNAME = "";
@@ -49,7 +51,7 @@ public class CerereEmitereAutorizatieFunctionarePDF extends HorizontalLayout
 
 	title = new Button("Generated PDF", VaadinIcon.FILE_PRESENTATION.create());
 	title.setEnabled(false);
-	// this.title.addStyleNames(new String[] { "borderless", "clearDisabled" });
+	
 	title.getClassNames().add("borderless");
 	title.getClassNames().add("clearDisabled");
 	titleLayout.add(title);
@@ -74,24 +76,7 @@ public class CerereEmitereAutorizatieFunctionarePDF extends HorizontalLayout
 
 	RouterLink routerLink = new RouterLink("", MainView.class);
 	routerLink.getElement().appendChild(backbtn.getElement());
-	this.backbtn.addClickListener(evt -> {
-	    if (fileName != null) {
 
-		File pd = new File(Utils.getFullPath(fileName, false));
-		try {
-		    System.out.println("before del: " + pd.getAbsolutePath());
-		    System.out.println("del1: " + Files.deleteIfExists(Path.of(pd.getAbsolutePath())));
-		} catch (IOException e) {
-
-		    e.printStackTrace();
-		}
-
-		PdfList.deleteFile(fileName);
-
-	    }
-	    RouteConfiguration.forSessionScope().removeRoute(CerereEmitereAutorizatieFunctionarePDF.class);
-	    RouteConfiguration.forSessionScope().removeRoute(NAME);
-	});
 	backLayout.add(routerLink);
 	content.add(backLayout);
 	content.setAlignItems(Alignment.CENTER);
@@ -100,10 +85,13 @@ public class CerereEmitereAutorizatieFunctionarePDF extends HorizontalLayout
 
 	setSizeFull();
 
-	UI.getCurrent().getPage().executeJs(
-		"window.addEventListener('beforeunload', function (e) {    $0.$server.windowClosed(); var nAgt = navigator.userAgent;if ((verOffset=nAgt.indexOf('Chrome'))!=-1) { (e || window.event).returnValue = null ; } return; });",
-		getElement());
-
+	 UI.getCurrent().getPage().executeJs(
+		 "window.addEventListener('beforeunload', () => $0.$server.windowClosed()); ",getElement()); //does not trigger on tab close !!!!!!!
+	 UI.getCurrent().getPage().executeJs(
+		 "window.addEventListener('unload', () => $0.$server.windowClosed()); ",getElement()); //does  trigger on tab close !!!!!!!
+	
+	
+	
     }
 
     @ClientCallable
@@ -113,17 +101,18 @@ public class CerereEmitereAutorizatieFunctionarePDF extends HorizontalLayout
 	try {
 	    System.out.println(Files.deleteIfExists(Path.of(Utils.getFullPath(fileName, false))));
 	} catch (IOException e) {
+
 	    e.printStackTrace();
 	}
-	PdfList.deleteFile(fileName);
+	if(PdfList.isFilePresent(fileName))
+	    PdfList.deleteFile(fileName);
 	RouteConfiguration.forSessionScope().removeRoute(NAME);
-	RouteConfiguration.forSessionScope().removeRoute(CerereEmitereAutorizatieFunctionarePDF.class);
+	RouteConfiguration.forSessionScope().removeRoute(AdeverintaRadiereAutoPDF.class);
     }
-
 
     @Override
     public void afterNavigation(AfterNavigationEvent event) {
-	Map<String, List<String>> parameters = event.getLocation().getQueryParameters().getParameters();
+	parameters = event.getLocation().getQueryParameters().getParameters();
 
 	if (fileName == null) {
 	    fileName = new String(parameters.get("tm").get(0));
@@ -132,9 +121,32 @@ public class CerereEmitereAutorizatieFunctionarePDF extends HorizontalLayout
 
 	    pdfView.add(Utils.getFullPath(fileName, true));
 	}
-
-	RouteConfiguration.forSessionScope().removeRoute(CerereEmitereAutorizatieFunctionarePDF.class);
+	
+	RouteConfiguration.forSessionScope().removeRoute(AdeverintaRadiereAutoPDF.class);
 	RouteConfiguration.forSessionScope().removeRoute(NAME);
+    }
 
+    @Override
+    public void beforeLeave(BeforeLeaveEvent event) {
+	
+
+	if (fileName == null) {
+	    fileName = new String(parameters.get("tm").get(0));
+	}
+	if (!fileName.isEmpty()) {
+
+	    pdfView.add(Utils.getFullPath(fileName, true));
+	}
+	try {
+	    System.out.println(Files.deleteIfExists(Path.of(Utils.getFullPath(fileName, false))));
+	    if(PdfList.isFilePresent(fileName))
+		    PdfList.deleteFile(fileName);
+	} catch (IOException e) {
+
+	    e.printStackTrace();
+	}
+	RouteConfiguration.forSessionScope().removeRoute(AdeverintaRadiereAutoPDF.class);
+	RouteConfiguration.forSessionScope().removeRoute(NAME);
+	
     }
 }
